+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.5.1
+
+* added examples
+* enhanced FVector
+* changed error calculation
+## 0.5.0
+
+* changed the constructor of TfannNetwork
+
 ## 0.4.0
 
 * many changes to api